[FIX] Autosave cmd conflict whit fuzzyfinder
The previous version would execute the autosave command whenever exiting InsertMode, however, this caused conflicts whe the file buffer was empty. For example, whe trying to execute the theme change command or Telescope from the file explorer. This would result in the autosave attempting to write to an empty buffer, causing errors. To solve this, a simple check was implemented to verify if the buffer is empty, and the autosave is only executed in the context of modifying a file